+++ CHANGES.rst
@@ -7,7 +7,8 @@ Changelog
 
 - *organizzazione_riferimento* field for Persona no more required.
   [cekk]
-
+- servizi_offerti in UO serializer now returns only related Servizi.
+  [cekk]
 
 2.0.6 (2021-04-16)
 ------------------