+++ CHANGES.rst
@@ -9,7 +9,10 @@ Changelog
   [cekk]
 - Now controlpanel settings entries can be multilanguage.
   [cekk]
-
+- *organizzazione_riferimento* field for Persona no more required.
+  [cekk]
+- servizi_offerti in UO serializer now returns only related Servizi.
+  [cekk]
 
 2.0.6 (2021-04-16)
 ------------------